+++ HPXMLtoOpenStudio/tests/test_validation.rb
@@ -76,6 +76,9 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
                             'dhw-frac-load-served' => ['Expected sum(FractionDHWLoadServed) to be 1 [context: /HPXML/Building/BuildingDetails, id: "MyBuilding"]'],
                             'dhw-invalid-ef-tank' => ['Expected EnergyFactor to be less than 1 [context: /HPXML/Building/BuildingDetails/Systems/WaterHeating/WaterHeatingSystem[WaterHeaterType="storage water heater"], id: "WaterHeatingSystem1"]'],
                             'dhw-invalid-uef-tank-heat-pump' => ['Expected UniformEnergyFactor to be greater than or equal to 1.45 [context: /HPXML/Building/BuildingDetails/Systems/WaterHeating/WaterHeatingSystem[WaterHeaterType="heat pump water heater"], id: "WaterHeatingSystem1"]'],
+                            'dhw-mixing-valve-setpoint-high' => ['Expected MixingValveSetpoint to be less than or equal to HotWaterTemperature'],
+                            'dhw-mixing-valve-setpoint-low' => ['Expected MixingValveSetpoint to be greater than or equal to 105 deg-F'],
+                            'dhw-setpoint-low' => ['Expected HotWaterTemperature to be greater than or equal to 105 deg-F'],
                             'dishwasher-location' => ['A location is specified as "garage" but no surfaces were found adjacent to this space type.'],
                             'duct-leakage-cfm25' => ["The value '-2.0' is less than the minimum value allowed",
                                                      "The value '-3.0' is less than the minimum value allowed"],
@@ -263,7 +266,7 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
                             'missing-elements' => ['Expected NumberofConditionedFloors [context: /HPXML/Building/BuildingDetails/BuildingSummary/BuildingConstruction, id: "MyBuilding"]',
                                                    'Expected ConditionedFloorArea [context: /HPXML/Building/BuildingDetails/BuildingSummary/BuildingConstruction, id: "MyBuilding"]'],
                             'missing-epw-filepath-and-zipcode' => ['Expected Site/Address/ZipCode or BuildingDetails/ClimateandRiskZones/WeatherStation/extension/EPWFilePath'],
-                            'missing-hpwh-containment-volume' => ['Expected HPWHContainmentVolume if HPWHInConfinedSpaceWithoutMitigation="true" [context: /HPXML/Building/BuildingDetails/Systems/WaterHeating/WaterHeatingSystem/extension[HPWHInConfinedSpaceWithoutMitigation="true"], id: "WaterHeatingSystem1"]'],
+                            'missing-hpwh-containment-volume' => ['Expected HPWHContainmentVolume if HPWHInConfinedSpaceWithoutMitigation="true" [context: /HPXML/Building/BuildingDetails/Systems/WaterHeating/WaterHeatingSystem[WaterHeaterType="heat pump water heater"]/extension[HPWHInConfinedSpaceWithoutMitigation="true"], id: "WaterHeatingSystem1"]'],
                             'missing-inverter-idref' => ['Expected AttachedToInverter if multiple Inverters are specified [context: /HPXML/Building/BuildingDetails/Systems/Photovoltaics/PVSystem[count(../Inverter) > 1], id: "PVSystem1"]',
                                                          'Expected AttachedToInverter if multiple Inverters are specified [context: /HPXML/Building/BuildingDetails/Systems/Photovoltaics/PVSystem[count(../Inverter) > 1], id: "PVSystem2"]'],
                             'missing-skylight-floor' => ['Expected ../../AttachedToFloor'],
@@ -327,6 +330,16 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
       when 'dhw-invalid-uef-tank-heat-pump'
         hpxml, hpxml_bldg = _create_hpxml('base-dhw-tank-heat-pump.xml')
         hpxml_bldg.water_heating_systems[0].uniform_energy_factor = 1.4
+      when 'dhw-mixing-valve-setpoint-high'
+        hpxml, hpxml_bldg = _create_hpxml('base.xml')
+        hpxml_bldg.water_heating_systems[0].temperature = 130
+        hpxml_bldg.water_heating_systems[0].mixing_valve_setpoint = 131
+      when 'dhw-mixing-valve-setpoint-low'
+        hpxml, hpxml_bldg = _create_hpxml('base.xml')
+        hpxml_bldg.water_heating_systems[0].mixing_valve_setpoint = 100
+      when 'dhw-setpoint-low'
+        hpxml, hpxml_bldg = _create_hpxml('base.xml')
+        hpxml_bldg.water_heating_systems[0].temperature = 100
       when 'dishwasher-location'
         hpxml, hpxml_bldg = _create_hpxml('base.xml')
         hpxml_bldg.dishwashers[0].location = HPXML::LocationGarage
@@ -1005,7 +1018,6 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
                                                          'EnergyFactor should typically be greater than or equal to 0.4.',
                                                          'EnergyFactor should typically be greater than or equal to 0.4.',
                                                          'No space cooling specified, the model will not include space cooling energy use.'],
-                              'dhw-setpoint-low' => ['Hot water setpoint should typically be greater than or equal to 110 deg-F.'],
                               'erv-atre-low' => ['Adjusted total recovery efficiency should typically be at least half of the adjusted sensible recovery efficiency.'],
                               'erv-tre-low' => ['Total recovery efficiency should typically be at least half of the sensible recovery efficiency.'],
                               'ev-charging-methods' => ['Electric vehicle charging was specified as both a PlugLoad and a Vehicle, the latter will be ignored.'],
@@ -1014,6 +1026,8 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
                               'heat-pump-defrost-backup' => ['BackupHeatingActiveDuringDefrost does not apply when system has separate backup heating'],
                               'heat-pump-low-backup-switchover-temp' => ['BackupHeatingSwitchoverTemperature is below 30 deg-F; this may result in significant unmet hours if the heat pump does not have sufficient capacity.'],
                               'heat-pump-low-backup-lockout-temp' => ['BackupHeatingLockoutTemperature is below 30 deg-F; this may result in significant unmet hours if the heat pump does not have sufficient capacity.'],
+                              'hpwh-120-dedicated-circuit-backup-capacity' => ['BackupHeatingCapacity should typically be less than 6824 Btu/hr for a HPWH on a 120V dedicated circuit.'],
+                              'hpwh-120-shared-circuit-backup-capacity' => ['BackupHeatingCapacity should typically be less than 3412 Btu/hr for a HPWH on a 120V shared circuit.'],
                               'hvac-dse-low' => ['Heating DSE should typically be greater than or equal to 0.5.',
                                                  'Cooling DSE should typically be greater than or equal to 0.5.'],
                               'hvac-capacities-low' => ['Heating capacity should typically be greater than or equal to 1000 Btu/hr.',
@@ -1115,9 +1129,6 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
             water_heating_system.energy_factor = 0.1
           end
         end
-      when 'dhw-setpoint-low'
-        hpxml, hpxml_bldg = _create_hpxml('base.xml')
-        hpxml_bldg.water_heating_systems[0].temperature = 100
       when 'erv-atre-low'
         hpxml, hpxml_bldg = _create_hpxml('base-mechvent-erv-atre-asre.xml')
         hpxml_bldg.ventilation_fans[0].total_recovery_efficiency_adjusted = 0.1
@@ -1142,6 +1153,12 @@ class HPXMLtoOpenStudioValidationTest < Minitest::Test
       when 'heat-pump-low-backup-lockout-temp'
         hpxml, hpxml_bldg = _create_hpxml('base-hvac-air-to-air-heat-pump-1-speed-lockout-temperatures.xml')
         hpxml_bldg.heat_pumps[0].backup_heating_lockout_temp = 25.0
+      when 'hpwh-120-dedicated-circuit-backup-capacity'
+        hpxml, hpxml_bldg = _create_hpxml('base-dhw-tank-heat-pump-120v-dedicated-circuit.xml')
+        hpxml_bldg.water_heating_systems[0].backup_heating_capacity = 7000
+      when 'hpwh-120-shared-circuit-backup-capacity'
+        hpxml, hpxml_bldg = _create_hpxml('base-dhw-tank-heat-pump-120v.xml')
+        hpxml_bldg.water_heating_systems[0].backup_heating_capacity = 3500
       when 'hvac-dse-low'
         hpxml, hpxml_bldg = _create_hpxml('base-hvac-dse.xml')
         hpxml_bldg.hvac_distributions[0].annual_heating_dse = 0.1